+++ packages/data-label/src/index.ts
@@ -1,0 +1,4 @@
+export * from './types';
+export * from './normalizeKey';
+export * from './buildLabelValues';
+export * from './composeLabel';